refactor(security-events): push and poll belong to SET delivery, not to Shared Signals
RFC 8935 and RFC 8936 describe how any Security Event Token travels; Shared Signals is one of
their users, not their owner. Half of that already sat in the right place - PollRequest,
PollResponse, DeliveryError, the media types - while the code using those models sat a package
above, so one concern was split across two packages with the usings crossing the boundary in one
direction only.

The push intake, the poll client, the event sink and the delivery result move down. What made the
intake framework-specific was three constructor parameters, nothing in its body: it now takes the
validation profile, what that profile expects, and the sink, and the consumer's own registration
binds them. That registration cannot name the profile with a keyed-service attribute, because an
attribute takes a compile-time constant while the key belongs to whoever registers - so it builds
the handler through the factory form, overriding exactly the two parameters it knows and resolving
the rest.

The poll client keeps only what RFC 8936 defines, a URL and a request. Reading that URL out of a
stream's delivery configuration is this framework's knowledge and stays here, as an extension
beside the receiver that uses it.

Back-channel logout does not become a consumer of any of this, and the move is not sold on that:
its request is a form-encoded parameter rather than a SET body, so its intake is its own. What the
move buys today is the layer; a second SET profile reusing the delivery is a consequence, not the
reason.

The replay identifier goes down with them - the push intake is one of its two callers, and its
rule is RFC 8417's rather than this framework's - and both suites of moved code go with the code.
