+++ src/Abblix.SecurityEvents/Delivery/ISecurityEventSink.cs
@@ -23,7 +23,7 @@
 using Abblix.SecurityEvents.Delivery;
 using Abblix.SecurityEvents.Validation;
 
-namespace Abblix.SharedSignals.Receiver.SecurityEvent;
+namespace Abblix.SecurityEvents.Delivery;
 
 /// <summary>
 /// Where validated events land: the application's half of a receiver, called once per accepted
